requirements.txt
# Polymarket 实时监控系统依赖

# 核心依赖
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
numpy>=1.24.0

# 日志
loguru>=0.7.2

# 重试机制
tenacity>=8.2.3

# 类型检查（可选，开发时使用）
mypy>=1.7.0
types-pyyaml>=6.0.12
types-python-dateutil>=2.8.19
